stream: track completed streams
In order to track active streams, we need to be able to detect when a
stream is complete, that is, when all sent data has been acknowledged
and when all incoming data has been read by the application.

We already have RecvBuf::is_fin() to detect completion on the receive
side. For the send-side, we need to track all ranges of sent data that
have been acknowledged by the peer.

When all data has been sent and acknowledged up to the stream's
send-side final size, then the stream is complete.

In the future we can use this to update local stream limits by sending
MAX_STREAMS frames, as well as to free memory when streams are not in
use anymore.

However, the use of RangeSet to track stream data might have an impact
on performance (particularly when packet loss generates gaps in the
data), so need to do more testing.